+++ tests/io/models/test_mlflow_model_logger_dataset.py
@@ -336,7 +336,6 @@ def test_pyfunc_flavor_python_model_save_and_load(
     model_config2 = model_config.copy()
     model_config2["config"]["run_id"] = current_run_id
     mlflow_model_ds2 = MlflowModelLoggerDataSet.from_config(**model_config2)
-    print(model_config)
 
     loaded_model = mlflow_model_ds2.load()